% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Area.R
\name{area_triangle}
\alias{area_triangle}
\title{Calculate the length of the area of an ellipse.}
\usage{
area_triangle(b, h)
}
\arguments{
\item{b}{The base of a triangle.}

\item{h}{The height of a triangle.}
}
\value{
The area of a triangle with specific values relating to its height and its base
}
\description{
Calculate the length of the area of an ellipse.
}
\examples{
Area_triangle (24 square cms)
Area_triangle (45 sqaure metres)
}
